refactor: PaymentClient interface로 생성
